FIR/UAST: commonize UMethod (entirely)
