--- uast/uast-kotlin-fir/testData/declaration/facade.comments.fir.txt
@@ -1,14 +1,0 @@
-KotlinUFile(
-    KotlinUClass(
-        FirKotlinUMethod(
-            UComment(// Single-line comment bound to fun foo)
-        )
-        FirKotlinUMethod(
-            UComment(/*
-             * Multi-line comment bound to extension fun buzz
-             */)
-            KotlinReceiverUParameter(
-            )
-        )
-    )
-)
