+++ uast/uast-kotlin-fir/testData/declaration/facade.comments.txt
